add: access to upload files by communes

+++ database/migrations/2020_09_29_235641_create_pro_programming_items_table.php
@@ -36,10 +36,10 @@ class CreateProProgrammingItemsTable extends Migration
             $table->decimal('workshop_session_time',5,2)->nullable();
 
             $table->string('professional')->nullable();
-            $table->integer('activity_performance')->nullable();
-            $table->decimal('hours_required_year',5,2)->nullable();
-            $table->decimal('hours_required_day',5,2)->nullable();
-            $table->decimal('direct_work_year',5,2)->nullable(); // Jornadas Directas Año
+            $table->decimal('activity_performance',5,1)->nullable();
+            $table->double('hours_required_year',5,2)->nullable();
+            $table->double('hours_required_day',5,2)->nullable();
+            $table->double('direct_work_year',5,2)->nullable(); // Jornadas Directas Año
             $table->double('direct_work_hour',15,8)->nullable(); // Jornadas Horas Directas Diarias double('column', 15, 8)
             $table->string('information_source')->nullable();
             $table->string('prap_financed')->nullable();
